Extract `TypeAnnotation` component into own file
